tests/wpt/web-platform-tests/presentation-api/controlling-ua/defaultRequest_success-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Presentation API, testing to start a new presentation with an default request setup. (success - manual)</title>
<link rel="author" title="Marius Wessel" href="http://www.fokus.fraunhofer.de">
<link rel="author" title="Tomoyuki Shimizu" href="https://github.com/tomoyukilabs">
<link rel="help" href="http://w3c.github.io/presentation-api/#dfn-controlling-user-agent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<p>Click the button or the menu item for presentation on your browser (for example, "Cast"), to start the manual test.</p>
<p>If your browser does not support <code>defaultRequest</code>, please click this button: <button id="notsupported">Not Supported</button>

<script>
    // disable the timeout function for the tests
    // and call 'done()' when the tests cases are finished.
    setup({explicit_timeout: true});

    // -------------------
    // defaultRequest init
    // -------------------
    var validUnixDate = new Date().getTime() + String(Math.floor(Math.random() * 1e5)),
            validPresURL = '../receiving-ua/idlharness.html#__castAppId__=2334D33A/__castClientId__=' + validUnixDate;

    navigator.presentation.defaultRequest = new PresentationRequest(validPresURL);

    // ------------------------------
    // Start New Presentation with
    // 'default request' Test - BEGIN
    // ------------------------------
    async_test(function(t) {
        navigator.presentation.defaultRequest.onconnectionavailable = t.step_func_done(function (evt) {
            var connection = evt.connection;

            assert_equals(connection.state, 'connecting', 'The initial state of the presentation connection is "connecting".');
            assert_true(!connection.id, 'The connection ID is set.');
            assert_true(typeof connection.id === 'string', 'The connection ID is a string.');
            assert_true(connection instanceof PresentationConnection, 'The connection is an instance of PresentationConnection.');
        });
        document.getElementById('notsupported').onclick = t.step_func_done(function() {
            assert_unreached('This browser does not support defaultRequest.');
        });
    }, '[Optional] The presentation was started successfully.');
    // ----------------------------
    // Start New Presentation with
    // 'default request' Test - END
    // ----------------------------
</script>
